ATTENTION: MyPlansUI still causes NullPointerException
Has to do with the workout log elements that compute the completion
percent.  Been working on fixing it and can't figure out how.   See
comment.

Also, there is a bug that recalculates the % complete of a plan element.
If you have a workout log item that it part of that plan element and you
go back and return, it doubles the %.